InfoChip: Re-name component to InfoChip
Originally we had considered handling the "info" and "filter"
chips as two variants of the same component. But we've decided
to separate them into InfoChip and FilterChip components
instead (styles will be shared where possible).

This change simply re-names all Chip stuff to InfoChip. FilterChip
has not been implemented yet but will be introduced in the near
future.

Bug: T323524
Change-Id: I0bca5dee5937eeba5afae5f875cec982f24c804c