mixins: Remove tokens import
Make callers import the tokens instead. Document this with warning boxes
on the Link and Icon pages, and in every example that uses a mixin.

This change is needed to allow mixins to work in MediaWiki, as well as
to allow other library consumers to choose which theme they use, rather
than being forced to use the wikimedia-ui theme.

Bug: T328602
Change-Id: Ie9de28afc608db338f2fc910bfc5c3a4d9ae0f9a