tokens: Move most deprecated tokens to separate aliases file
Move all deprecated tokens that are just aliases for other
non-deprecated tokens (or for a combination thereof) to a separate
aliases file. These aliases only exist for WikimediaUI Base's benefit,
so this change moves them to a separate set of output files that will
only be used by WikimediaUI Base.

Keep visualizing these deprecated aliases on the docs site, but move
them to a separate section at the bottom of each page.

This reduces the number of deprecated tokens in the main token set
from 98 to 9. The 9 remaining deprecated tokens have no
straightforward equivalent tokens that can be used instead, and we
want to keep their values here in Codex, rather than having those
live in WikimediaUI Base.

(`width-breakpoint-mobile` could have been moved to the aliases set,
but this change keeps it in the main set for consistency with the
rest of the `width-breakpoint` category.)

Bug: T330694
Change-Id: I771877f53305480c6490e0b8becf71682794905a