docs: Update token table style and add caption
Updating table style with outcome of design conversations.
Making the table easier to read and navigate.
Adding a visually hidden caption to the table to make it
more accessible. For this also add a new visually hidden class, an
equivalent to Codex Less mixin to VitePress' 'custom.css'.
Seems to be the more appropriate place than using the Codex one here.

Change-Id: Idc804281574dcf89e09da3e2a2af8f1c23d6ca48