build: Remove 'dist/' from import path for codex-icon-paths.less
Importing this file from the codex-icons package looked like this:
@import '@wikimedia/codex-icons/dist/codex-icon-paths.less';

We removed the 'dist/' part from the import paths for tokens and mixins
in 5b397293c86b61c, but we never did it for this .less file in the icons
package. This change does that, so that 'dist/' is now optional for all
Less imports from all of our packages.

Bug: T328602
Change-Id: Idb327a1cdacfd84168b0fc3dcb8476ec08908837