build: Don't delete built icon paths file in postpublish
The postpublish command for the icons package was deleting the
codex-icon-paths.less file from the dist directory, which is wrong and
causes issues when running the publish command twice in a row (for
example when doing a dry-run followed by a real publish). Instead, it
should delete the temporary copy of codex-icon-paths.less in the
codex-icons directory.

Change-Id: I3f9eed2a62094fb514cce0ba66800ad76ce27ecc