Simplify Codex build process via a JS build script
Introduces a `generateCodexBundle` function which contains
the logic for producing all the variants we need for a given
Codex library build (base, RTL, legacy, and legacy RTL). This
script takes a bundle name and a Vite UserConfig object, and
makes multiple calls to the Vite `build` method to generate
all necessary files. This function currently lives in
build/utils.mjs in the Codex workspace, but in the future we
may move it elsewhere.

Also adds a `build.mjs` file in the Codex workspace which
orchestrates the build process using Vite's JS API and this
new generateCodexBundle method. A similar script has been
added to `codex-docs` (to build the demo bundles) and to
`codex-search` (to build the Typeahead bundles).

Build-specific code has been removed from the vite.config.ts
files; these files are now only needed for development mode.

Potential future improvements: We may want to the following
in follow-up patches (they should be considered out-of-scope
for the patch here):

* Remove the `codex-search` package entirely in favor of
  generating that build from the main `codex` workspace;
  this is a breaking change and would also involve unpublishing
  or archiving the `codex-search` NPM module
* Move the `generateCodexBundle` script and any related shared
  utils or configs into a new "build" or "utils" package so that
  other workspaces within the monorepo can explicitly depend
  on this. Currently there are some cross-workspace imports
  happening that are easy to miss; however this was also the
  case before this patch, so things are not any worse than they
  already were on this account.

Bug: T335323
Change-Id: I4bcfc180ccbc389d50f4545781f0fa85b873707d