Icon: Support icons in CSS-only buttons in Chrome
Chrome supports the mask-image feature, but requires the -webkit
prefix. Since the CSS-only icon mixin may be used in places where
it is not run through autoprefixer, we need to add the -webkit
version of mask-image to the actual use of the property and to
all @supports queries related to it.

Bug: T335430
Change-Id: Ie7c1912b5abdf059d5679cff55cfc267befb19d0