styles: Add mixins for flush layouts
This patch adds a mixin for quiet buttons that sets a
negative margin equal to the horizontal padding of the
button, making the content of the button appear "flush"
with the edge of the UI.

We have chosen not to document this mixin because we
would like to reevaluate the design decisions behind it
and potentially create a new button variant (e.g. a
quiet button with no background change on hover) or
dropdown menu button component in the future.

In a subsequent patch, we will add a media query for icon-only
buttons that applies a different margin at different viewport
widths.

Bug: T322436
Change-Id: Idf681b70119f4379dbbefa0f27d762971b667c8c