Tabs: Add CSS-only version
This patch adds an implementation of CSS-only tabs that requires
some logic on the server-side to set ARIA attributes that the
Codex styles will use to set appropriate tab list styles and
display the current tab. This implementation ensures that nothing
else on the page will interfere with the current tab selection
(due to the unique URL query param) and a link to the page with
a specific tab being shown can be shared.

Bug: T325105
Change-Id: I0586a2c09c2ac5d5b5dd9c5ca03fab1cfaf13709