Tabs: Remove getLabelClasses, use aria-* for styling
Part of the change that enabled Tabs to support CSS-only usage
involved updating some of the component styles to rely on aria-*
attributes, since Vue-managed modifier classes might not be
available in a no-JS environment.

We actually don't need the modifier classes (--enabled, --disabled,
--active) at all now that this has been done. This change updates
the styles so that both the Vue and CSS versions of the component
rely on aria-* attributes for styling. The getLabelClasses() method
which applied these classes is no longer necessary and has been
removed.

Change-Id: I327f39b10a16340fe502ba55035b075a108dc8e2