build, styles: Enable `*width*` and `*height*` declaration strict value
Enabling
- 'min-width'
- 'min-height'
- 'width'
- 'height'
- 'max-width'
- 'max-height'
strict value rules for linting stylesheet properties.

Recent volunteer addition of Accordion has shown that those properties
are in need to be automatically linted against. We're already providing
a sufficient choice of tokens.

Note, that `fit-content` as general allowed value is a compromise, it's
debatable if it's better off in base tokens.

Also note, that
- Checkbox indeterminate line switches to absolute 1px value as it's
  more theme friendly relating to expectations
- I'm aligning the Message demos to use a size equivalent static value,
  we could replace them in future with equal value `min-size` token
- I'm updating the demo dialogs to use a standard token and default
  `padding`, it's still custom enough while not providing a weird
  vertical offset

Bug: T337279
Change-Id: I2343adf393a26d807e0671c98b9cfe646a3e49df