Tabs, docs: Don't show Vue code in the CSS-only HTML example
This is confusing for end users. Instead, show plain HTML, with the
attribute values that depend on the current tab substituted. Vitepress
recently introduced a feature where you can use Vue substitutions in
code snippets by adding -vue to the language code; make use of that.

See https://vitepress.dev/guide/using-vue#unescape-in-code-blocks for
documentation of this new VitePress feature.

Change-Id: I7eaa3deaf85428f92ed95fbfe8e8a3ef7094a32d