Dialog: Simplify padding styles
Always set the same amount of padding on the header, body and footer
regardless of whether a custom header/footer is being used or whether
dividers are being displayed.

- Header: 16px top/bottom, 24px left/right
- Body: 16px top/bottom, 24px left/right
    - If no header present: top changes to 24px
    - If no footer present: bottom changes to 24px
- Footer: 16px top, 24px bottom, 24px left/right

This replaces the following, much more complex padding logic:
- Header: 16px top, 32px bottom, 24px left/right
    - If custom header: no padding
    - When dividers are shown: bottom changes to 16px
- Body: no top/bottom padding, 24px left/right
    - If custom header: top changes to 16px
    - If custom footer: bottom changes to 16px
    - When dividers are shown: top/bottom change to 16px
    - If no header present: top changes to 24px
    - If no footer present: bottom changes to 24px
- Footer: 32px top, 24px bottom, 24px left/right
    - If custom footer: no padding
    - When dividers are show: top changes to 16px

Custom headers/footers can still override the padding if they need to,
but this way they get a reasonable amount of padding by default, rather
than zero.

This change moves some of the padding from the header/footer to the
body in the no-dividers case, but this causes no visual change.

Also remove the now-unused --has-custom-header and --has-custom-footer
CSS classes.

Change-Id: I84ab2f18e7e77db99c5935040828e38b73e2cd29