Button, styles: Fix bare :not() selector
This selector was using an ampersand at the top level, which doesn't
work but also doesn't cause a warning. This didn't generate any
dangerously underspecified CSS with the current code, because every
selector in its block specified a .cdx-button-foo class name, but it
would have been very easy to accidentally generate such selectors in the
future.

Change-Id: Ic867e7c834a73b9893c2dec6884a2d595445bfb2