styles, Button, Message: Remove some `:not()` selectors
Button and Message both have props with multiple possible values:
- Button has weight and action, each of which have 3 options
- Message has type, which has 4 options

The code used to employ `:not()` selectors to enable us to set
styles for the default value of these props not by using the
associated CSS class but by eliminating all of the non-default
classes. This means that the user does not need to include the
default classes when building the markup for a CSS-only version.

However, the chains of `:not()` selectors were very specific,
which caused issues of performance and unexpected visual
regressions. This patch refactors Button and Message styles in
the following ways.

Button:
- Styles for normal weight, default action, enabled buttons are
  now applied to the `.cdx-button:enabled` selector. They can
  then be overridden for different weights and actions
- Styles for enabled buttons of all weights and types are now
  nested under that `.cdx-button:enabled` selector
- Styles for all disabled buttons appear below that. Disabled
  styles are the same for all buttons except quiet weight, which
  has a few overrides
- Disabled stylelint's `max-nesting-depth` rule. Because all
  styles are now nested under the root element, and there are
  so many variations, a lot of nesting is inevitable.

Message:
- Styles for block, notice type messages are now the default.
  They are overridden for messages of other types and for
  inline messages.
- Generally reorganized styles for better readability

Bug: T339141
Change-Id: Ief2a3f945bc9419c299ab43b9b2c7035d7b54da0