build: Don't make stylelint require tokens where they don't exist
We don't have tokens for text-align, caption-side or for the
list-style-* properties, so every time we used these properties we had
to override the stylelint warning about needing to use tokens here.
Instead, comment out these properties from the list of properties where
tokens are enforced, until such time as we have tokens for these.

Change-Id: I2e6b57a6316047c7d046aec9a0518b424cf9bfd1