Icon: Make <use> work in CSS-only icons
Some of our icons contain tags like <use xlink:href="#foo">. These icons
were broken as CSS-only icons for two reasons.

First, the # in the xlink:href attribute was put in the data URI
unencoded, terminating the URI (the text after the # is interpreted as
the fragment). To fix this, make buildIconsLess.js encode all #
characters in the SVG. All other characters with special meanings in
URIs are allowed in the data portion of data URIs.

Second, the xlink:href attribute only works if the xlink namespace is
added to the <svg> tag, and we didn't do that. Add it in all the places
in css-icon.less where we generate <svg> tags, and also in Icon.vue for
completeness (though it wasn't broken there).

The xlink:href attribute is deprecated, and the href attribute is
recommended instead. We could switch to this and not need the xlink
namespace, but unfortunately href doesn't work in some of the older
browsers we support for CSS-only icons (Chrome 31-49, Safari 9.1-12,
Opera 18-36), so we'll have to keep using xlink:href until such time as
we drop support for those browsers.

Bug: T340570
Change-Id: Ib52bd4caa6d5f9c0d159737e7dd057a72caa331c