Checkbox, Radio, ToggleSwitch: Use Label internally
This patch updates the 3 binary input components to use the Label
component internally. For all 3 components, this means adding
the ability to include a description that is semantically tied
to the input. For ToggleSwitch, we also allow for a visually-
hidden label.

Note that for all 3 components, the old CSS-only implementations
still work (the most important aspect here is that, when you hover
over a label, the input should show hover styles).

- Update the 3 components to use Label instead of a <label> element.
  In the rare case that the default slot is not provided (such as
  when `aria-label` is used), the Label component will not render.
- Add a description slot to each component. This will only be
  passed to the Label component if slot content is provided.
- Add a `hideLabel` prop to ToggleSwitch that gets passed into
  label as `visuallyHidden`
- Update snapshots and tests accordingly
- Update styles to ensure that hovering over the label produces
  the proper input styles for both the new markup and for the
  old CSS-only implementations with simple <label> elements
- Add examples of the new features for both Vue and CSS-only
  implementations

Bug: T339013
Change-Id: Ia3a41fe763111f6e803e979fabc5248905e63111