Field: Improve reactivity for input ID and description ID
These two IDs were provide()d conditionally in the setup() function,
based on the value of the isFieldset prop and the presence of the
description slot *at setup time*. This breaks reactivity, because if
the prop or slot change later, the setup() function is not rerun and
things that should have been provide()d aren't.

In practice, this meant that a Field that started out with
:is-fieldset="true" but changed to false later would not assign an ID to
the input inside it (even though the <label>'s for attribute would point
to that ID), and would not set aria-describedby on the input.
Conversely, a Field that starts out as a non-fieldset but changes to a
fieldset will set an aria-describedby on the input that points to a
now-missing ID. (The latter can be tested in the configurable demo.)

This change fixes this issue by unconditionally provide()ing the input
ID and descripton ID, and making the provided values computed refs that
change based on the prop/slot, rather than static strings.

Change-Id: Ie466e38917b54c15e940fdb4768103e1ac67aa18