Label, Field: Do not render description unless it exists
Currently, an empty span is rendered for the description whether
there is slot content or not. This changes the Label component
to conditionally render the span only if description slot content
is provided. A similar update must be made in the Field component
to only pass in its description slot to Label if there is slot
content.

Change-Id: I6a2a9d306da4c986861957cef49f13fd3ec36360