Radio: Add focus styles for unchecked radios
We previously assumed that if a Radio is focused, it is also checked.
However, this doesn't account for a radio group where nothing is
checked initially, and the user needs to be able to tab into the
group and perceive that they are focused on a radio.

This patch adds a focus outline (actually a border); the same one
applied to a focused, unchecked Checkbox. It also removes the
initial selection from the "inline radios" demos (Vue and CSS-only)
for testing purposes and better demonstration to users.

Finally, a new demo was added to show a radio group with no
initial selection.

Bug: T343219
Change-Id: I5c6bff2f94683f6dbfc07469b2454605eb6fae19