ToggleSwitch: Use role="switch"
The switch role is meant for circumstances where "on" and "off" makes
more sense semantically than "checked" and "unchecked." This patch
adds the switch role to the ToggleSwitch's internal checkbox input,
which will improve the experience when using a screen reader.

Previously, a screen reader user would head "unchecked checkbox" or
"checked checkbox" for a ToggleSwtich. Now, they will hear "switch
on" or "switch off."

This patch also updates the docblock comment of ToggleSwitch to remove
outdated information about inclusion of a label and to add a sentence
about the semantic purpose of a switch.

Bug: T342396
Change-Id: Ide19100ebe19cf6c0fa6b030e01e956e0d661566