Menu: Always clear the highlighted state when the menu closes
We were previously only doing this if nothing was selected. In practice,
this meant that selecting an item with the keyboard would result in the
highlighted state still being set after the menu had closed. Selecting
an item with the mouse would not, because the menu closing caused a
mouseleave event on the highlighted item, which un-highlighted it.

Also make the expanded watcher more readable by grouping the "on open"
code and "on close" code in two separate branches; these two shouldn't
have any overlap.

Clean up the code further by making the second parameter to
handleMenuItemChange() required, and explicitly passing null (rather
than undefined) when unsetting things.

Change-Id: I05bb9a45e3db7659c75c87aa6654a163d08ae9af