Menu, Select: Allow keyboard navigation through typing
When the user types a character while a Select is focused, highlight the
first item whose value begins with that character. If the user types
multiple characters in rapid succession (less than 1.5 seconds apart),
highlight the first item that begins with that string. For example, in a
dropdown with a list of countries, typing "c" might highlight
"Cambodia", but typing "co" might highlight "Colombia".

This approach is similar to the APG example:
https://www.w3.org/WAI/ARIA/apg/patterns/combobox/examples/combobox-select-only/
and is also used in OOUI.

Both the APG example and OOUI have a feature where repeating the same
character cycles through all matches for that character. This patch
implements that as well. This means pressing "c" highlights "Cambodia",
pressing "c" again highlights "Cameroon", pressing "c" a third time
highlights "Canada", etc.

One thing we might want to consider adding to this in the future is
allowing menu items to specify a "keyboard label" (or whatever we call
it, that's probably not the best name), which is used instead of the
displayed label to match typed characters.

This is technically a breaking change to the delegateKeyNavigation
method, but I wasn't able to find any code outside Codex using this
method.

Bug: T343490
Change-Id: Ib1ef82945b01fc5592c2bfa6326b675f2237fb64