Icon: Remove click handler
Because the Icon component has a click handler and event, it is
viewed as "clickable" by assistive technology and must have
"interactive semantics." Typically, icons should not be interactive
and a button should be used if click interaction is needed.

This patch removes the click handler and event. I have confirmed
that the following instances of icons with click handlers still
work:
- Clearable TextInput
- WikiLove dialog close icon
- WikiLambda function tester table info icon

Bug: T329321
Change-Id: I661042a6dad453c77f30c94f51e3de95624f16aa