code: Add component name to warnings
When we emit a warning via Vue's warn() method, it may be unclear
whether the warning is coming from Vue or from Codex. This patch
adds the component name as a prefix to all existing warnings so
it will be clear that the warning comes from Codex and which
component it relates to.

Change-Id: I2a8c34d3403cff95471e5dbd7c7369df16f38051