Menu: Don't set aria-activedescendant when the menu is closed
aria-activedescendant is managed by the parent component, which relies
on Menu's getHighlightedMenuItem() method. Sometimes there can be a
highlighted item even though the menu is closed (this is a bug, and will
be fixed in a future change), resulting in aria-activedescendant
pointing to an invisible element.

To prevent this from happening, make getHighlightedItem() return null if
the menu is closed.

Change-Id: I7d00bb1ede23128637c5e75e4bef1f7511c45ae9