docs, Button: Add docs on role="button" for fake buttons
We allow button styles to be applied to non-button elements and
demonstrate this in the CSS-only section of the Button page and
in our internal Sandbox. This patch adds documentation on when
to use `role="button"` and adds the role to the checkbox hack
demo in the Sandbox (since those fake buttons do trigger an
action).

Bug: T342295
Change-Id: I6940e494ed4b5d9860dfe2e56fa4bbad7659ac9a