Tabs: Override browser default styles for <button>
We recently changed tabs from <li>s to <button>s, but we didn't override
the background and border properties from the browser's default button
styles. We didn't realize this because VitePress provides similar
overrides for all buttons, hiding the issue. (The issue was visible in
the sandbox, but we didn't look there, so we only found it when running
visual regression testing in MediaWiki after the 0.16.0 release.)

Change-Id: If21e3eccdb5544aed11bca4febded8c8ba400d2f