Menu: Simplify logic for highlighting selected item
We needlessly had this logic in two places, both in the watcher that
fires when the menu is opened, and in the key handler when opening the
menu in response to e.g. an Enter key press. The latter is unnecessary,
so remove it. This simplifies the key handler code a bit.

Change-Id: I210b8b69b509c7c2c613f9b8601a67202a2fac5a