docs: Add MediaWiki versions of inlined Less examples
For every example in the docs where we demonstrate Less code that
imports the tokens and/or the mixins, provide a MediaWiki version as
well. This mostly affects examples of CSS-only components, but also
affects a few examples on the usage page.

Generated by replacing the following regex using VSCode:

```less
((?:// .*
)?@import .*(?:
@import .*)?)
((?:.|
)+?)
```

with:

:::code-group

```less [NPM]
$1
$2
```

```less [MediaWiki]
@import 'mediawiki.skin.variables.less';
$2
```

:::

Change-Id: I7aa4f84b879f09a1cf89db68b00a3c5136695180