docs: Link to the appropriate path in the version banner
On the main and patch preview sites, a banner tells you that you are not
looking at the official docs, and provides a link to the official docs.
This patch updates that link to include the current path, so the user
is taken to the equivalent of the page they were on.

Change-Id: I0edff4378a2b79288c1f5ba2798aa7153a04cd71