[Breaking] Dialog: incorporate Vue's <teleport> feature
Updates the Dialog component to rely on Vue's <teleport> feature
internally.  The component now accepts a "target" property which is
passed to the "to" prop of the <teleport>.

Additionally, a "renderInPlace" prop has been added that can be used
to opt-out of teleport behavior.

If no target is provided and renderInPlace is not set to true, then
Dialog components will teleport to the end of the body tag by default.
However, the default location can be changed by providing a value
for "CdxTeleportTarget" somewhere higher up in the component hierarchy.

In either case, manually specifying a teleport target will always
take priority over any default value.

This patch also makes some modifications to the Docs site (which
uses SSR) to make sure that teleport-based Dialogs work properly
there. Dialog components in demos have all been wrapped in Vite's
custom <client-only> component, and the Wrapper component now
provides a local teleport target which ensures that direction-specific
styles for demos are preserved.

Bug: T343476
Change-Id: I1b64201e8f64664c6bebb068c578fa9d6347dd1a