FilterChipInput: Ensure error state is unset
When you try to add a FilterChip with the same text as an
existing FilterChip, an error state will be set. Currently,
if you remove the duplicate FilterChip, the error state
will perist.

This patch watches the inputChips prop and, when it updates,
checks for duplicate values between the chips and the input.
The status is set based on that test.

Change-Id: I17ce538c4b18c068ed192773f939f687a7349eb0