FilterChipInput: Make chips editable; add chip on blur
Allows the user to edit the text of an existing chip in a
FilterChipInput by clicking on it. This action converts the
chip back into regular text inside the input (or the separate
input element, if that option has been enabled). When the user
submits the text in the input, it is turned back into a chip.

Also, on blur, if there is value in the input, it should be
added as a new chip.

To achieve these two things, the following changes were made:
- Add a new addChip() method that first checks for duplicates,
  then adds the current input value as a new chip (if there is
  an input value) and clears the input
- Run addChip() on input blur
- When you click on a FilterChip in a FilterChipInput, remove the
  chip but add its value to the input. Before this happens, the
  blur handler has already run addChip(), which adds any current
  input text as a new chip

Additionally, when a chip is removed, focus is placed back on
the input to help clarify where focus is after chip removal.

Bug: T343474
Change-Id: I30f2c923928779589e1a1ac789d74c652f5f040e