docs: Fix "Copy code" button, make it work with code groups
The "Copy code" button stopped working; for whatever reason,
slots.code()[0].el no longer gives us the rendered HTML element.
Instead, obtain it through a template ref.

But that's not enough. Since code groups were introduced, the text in
the code sample area can change without the Wrapper component being
notified about it (when the user switches between tabs in the code
group). To address this, pass a DOM node instead of a string to
CopyTextButton, and extract the text only when the user clicks the copy
button.

Bug: T342403
Change-Id: I39b73fe3a6f1e784f2cdacd7110553d19bdd1353