Label: Add CSS-only version
In addition to adding docs on CSS-only usage, a change was
made to the Label component itself: instead of relying on a
class, .cdx-label--is-legend, to unset <legend> styles, just
use the legend element as a selector. This is what we are
doing in the Field component to unset fieldset styles, and
saves CSS-only users one class to include in their markup.

Bug: T339021
Change-Id: I6ac484ec52f0a6f06890d60d5d647cca7a2c1b2f