Dialog: Set aria-hidden and inert on the rest of the page when open
When the dialog is opened, set the aria-hidden and inert attributes on
every element on the page that is not an ancestor of the dialog (or
rather, only on the siblings of the ancestors of the dialog, as that
covers everything). Unset these when the dialog is closed.

Also ensure that the cleanup steps that run when the dialog is closed
also run if the dialog is unmounted without being closed.

Bug: T344178
Change-Id: I372040a358d9713ab0fa37868afdcef8a2b383c5