TextInput, Combobox: Set consistent min-width regardless of icons
Combobox uses a TextInput at its minimum width, and then reduces that
min-width by the width of the expander button, so that the width of a
Combobox is the same as the width of other components (e.g. Select).
However, TextInput set its min-width on the <input>, which doesn't
account for start/end icons and the clear button. This caused Comboboxes
with a start/end icon to be wider than the expected min-width, and also
caused jarring behavior where a clearable Combobox would become wider
when text was typed into it, because the clear button appeared (and
narrower again when the text was removed and the clear button
disappeared).

To address these issues, this change:
- Sets TextInput's min-width on the entire component instead of the
  <input>, so that it also captures the clear button and start/end icons
- Moves up Combobox's min-width override accordingly
- Sets size="1" on TextInput's <input>, to make the min-width behavior
  work correctly. Without this, the size attribute defaults to 20, which
  causes the browser to set an implicit min-width of 207px. This
  prevents us from setting a consistent min-width on TextInputs and
  Comboboxes, because in some cases that requires reducing the width of
  the <input> to less than 207px (when more than two of start icon, end
  icon, clear button, and Combobox expander button are present)

Bug: T342313
Change-Id: I874bf50c0714859948c22a5cfe2e4c88f8eca614