TextInput: Fix padding when both clear button and end icon appear
When both the clear icon and end icon were shown, we were setting 1px
too much padding. The spacing-horizontal-input-text-two-end-icons token
incorrectly included border-width-base (1px) to compensate for the
input's right border. But that should only be done when positining
(setting the right: property), not for padding.

Change the spacing-horizontal-input-text-two-end-icons token to remove
border-width-base from its computation, and add it back in where we're
using this token for positioning (but not where we're using it for
padding).

Change-Id: I7ee494a7e6e71fbd2036b2ea24fe90b3aa65b6f5