styles, Field: Unset margin on the root element
We're already unsetting border and padding since browsers add
styles for those properties for the <fieldset> element. In
MediaWiki, and potentially other places, margin styles are also
set on the <fieldset> element. In practice, this means that
margins appear on multi-input fields but not single-input fields
in MediaWiki, which is confusing when trying to apply layout
styles in a form.

This patch unsets the margin on the root element of the Field
component to ensure consistent styles for all fields. The
margin can easily be overridden in feature code to add space
between form fields. We may want to do the same in a future
Form component: all child Fields can have a margin-bottom
applied, for example.

Change-Id: I6e1d91d1ed27a799b9ec056e327972f528ca23b7