useI18n: Add basic i18n system
This adds a basic i18n system consisting of a useI18n() composable that
components can use for i18n messages. Components must pass in a symbolic
name for the message, and a hard-coded English default.

Software that uses Codex can use `provide( 'CdxI18nFunction', ... );` to
provide a function that takes a message key and parameters and returns a
translated string. The composable invokes this function, and uses its
return value if there is one. If no function was provided, or if the
function returns null or undefined, it falls back to the hard-coded
default value.

In the future, we may wish to improve this system by centralizing the
list of message keys and their defaults.

Bug: T368442
Change-Id: I6d1bda08b5f0367fa8ad04266d3cfe15ed7ebe1c