ChipInput: Add input text as chip when focus leaves component, not input
When the input was blurred, we added the text in the input as a new
chip. However, using the arrow keys to move the focus from the input to
one of the chips also counts as blurring the input, and we don't want to
add the text as a chip in that case.

Instead, listen for the focusout event on the root element, and add a
new chip when a focusout event fires that indicates the focus is moving
to another element that is not inside the component.

This also requires adding an explicit addChip() call to
handleChipClick(), because this can no longer lean on the input blur
handler taking care of that. The addChip+removeChip combination only
works if we wait for nextTick in between, which is a bit hacky; I'll
propose a follow-up task that cleans this up, but I wanted to keep this
patch simple.

Bug: T344848
Change-Id: Ib8c2d3fad75d640e1b0b54c5f55d842f459d1523