build: Prefer CJS build to UMD for most usage
Updates the various build scripts to produce CommonJS builds
alongside the ESM bundles, and updates package.json files
as well as various docs and scripts to rely on the CommonJS
files for most things.

A UMD bundle is still produced as part of the main Codex
build process, but this is only intended for users who
download Codex via a CDN otherwise use it via plain script-tag
inclusion.

Bug: T345688
Change-Id: Idfce607f3b34a2126b1d0a196934f034ff155aaa