build: Make the build config the primary tsconfig.json file
In the codex and codex-icons packages we had two tsconfig files: one for
building the .d.ts files named tsconfig-build.json, and one for type
checking named tsconfig.json. The type checking file extends the build
file, and adds files to the "include" list that we don't build type
definitions for.

Keep this setup, but make the build config the primary config file, and
the type checking config the secondary one:
- Rename tsconfig.json -> tsconfig-check.json
- Rename tsconfig-build.json -> tsconfig.json

This means the build config is now the primary tsconfig that Vite will
look at to determine how it should transform TypeScript in its build
process, which makes more sense. This also frees us up to change
tsconfig-check.json to use features that we wouldn't want to get into
the build output.

Specifically:
- Change tsconfig-check.json to set "target" to "esnext". This enables
  top-level await
- Use top-level await in Codex's build.mjs, resolving a TODO comment

Change-Id: Id630aed260d859e7ff54e516f167a78d2039f3f3