build: Add SVGO & svglint to 'codex-docs'
This adds SVGO and svglint to the 'codex-docs' package build process.
It also optimizes the SVGs in the 'docs/assets' directory, in preface
of adding a huge number of SVGs for the component documentation, that
are supposed to follow MediaWiki's coding conventions.

Note that this is currently not requiring `title` elements as part of
the documentation assets. They are always featured with `alt`
attributes, we might add requirement for `title` post-MVP.

Bug: T347906
Change-Id: I6af55f5a515e7f9b4ebacb87dd5c5cf51101a871