docs: Add existing component guidelines image assets
This commit includes the first batch for the component guidelines of
- Links
- Buttons
- Button groups
- Checkboxes
- Radio buttons
- Toggle switches
- Selects
- Comboboxes
- Text inputs
- Messages
- Dialogs

for all moved guidelines. There will be a second batch for all new ones.
All of those are SVG linted and SVGO optimized.
Note, that those SVGs currently don't feature a `title` element, as
the benefit is not worth the effort for the MVP. The SVGs will all
already be featuring `alt` text in HTML. The titles will be added
after the 1.0 release.

Bug: T347906
Change-Id: Ib3a19f1f52251a153316ac9dc1f4f9afcef00a29