docs: Add new component guidelines image assets
This commit includes the second batch for the component guidelines of
- Accordion
- Card
- ChipInput
- Field
- Icon
- InfoChip
- Label
- Lookup
- Menu
- MenuItem
- ProgressBar
- SearchInput
- Tabs
- Tab
- TextArea
- Thumbnail
- ToggleButton
- ToggleButtonGroup
- TypeaheadSearch

for all new guidelines. There is also another batch for all moved ones.
All of these are SVG linted and SVGO optimized.
Note, that these SVGs currently don't feature a `title` element, as
the benefit is not worth the effort for the MVP. The SVGs will all
already be featuring `alt` text in HTML. The titles will be added
after the 1.0 release.

Bug: T347906
Change-Id: I4c1df00653ea6ad3b9b9fc74253790847517baab