MenuItem: Ensure proper hyphenation and wrapping of long words
We have an example of a MenuItem with a label that is a single,
very long German word. Long labels should be hyphenated and
wrap to a new line, but this wasn't working in Firefox or
Safari: Long capitalized words are intenionally not hyphenated
since they may be proper nouns, except in German[1]. Setting the
lang of the label to "de" fixed this in our example.

In case this might happen for other languages, this patch adds
`max-width: 100%` rules to the text elements to ensure that
long words will wrap instead of overflowing their container,
even if they aren't hyphenated.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=656879

Bug: T338814
Change-Id: I950d761611515834b9c4c3177622f61d9e938ac8