styles: Replace falsely applied `font-size-base` with `font-size-medium`
This slipped under the radar. The `font-size-base` token is only a magic
calculation helper token and should not be used directly. The correct
one (with user zoomable, accessible) relative size is
`font-size-medium`.
We'll care about making the documentation clearer in a follow-up, but
this is an easy and early fix for the direct a11y hurdle.
Note, that there's also an open question if we continue with the same
size for custom headings or if they should remain relative size and we
only override the default heading level for following Codex design.

Bug: T350346
Change-Id: I320af52518534c1ffd152dadeed68a1849fcf8d5