TextInput: Move border-radius from `<input>` to root element
Instead of applying a border-radius to the `<input>` element, apply it
to the `<div>` that wraps it.

This ensures that the focus outline matches the input in terms of which
corners are rounded. Previously, if a style override from the parent
component made one of the corners straight (set the border-radius to 0
for that corner), the focus outline would still be rounded at that
corner; with this patch, the focus outline is also straight at that
corner.

Moving the border-radius to the root element is also needed to allow
useFloatingMenu to change the border-radius on a TextInput when its
associated Menu flips from below to above (in a future patch).

Remove the element-with-menu-expanded mixin, and instead use the
--expanded classes on the parent element to override the `border-radius`
on the input when the menu is open.

Change-Id: Icac7f44af2e76a1dabefc956c058e7a1b54b74c6