docs: Adjust sidebar outline marker size
Currently we extend the base VitePress theme to enlarge the font size
of sidebar navigational items. This causes the outline marker that
indicates which navigation item is currently in the viewport to be
misaligned.

VitePress uses a pixel value for the height property, see:
https://github.com/vuejs/vitepress/blob/a482611d17197a0b7afc403891cd95f344e7a55f/src/client/theme-default/components/VPDocAsideOutline.vue#L74

Instead, compute the height property using the same design tokens that
are used for the navigational items, so that the outline marker has the
same height as each navigational item.

Bug: T350045
Change-Id: I01e037075f567896072ac3647260abfaa411ce44